--- packages/components/src/molecules/VirtualizedValidation/index.ts
@@ -1,1 +1,0 @@
-export {default as VirtualizedValidation} from './VirtualizedValidation';
