+++ packages/components/src/molecules/index.ts
@@ -8,6 +8,5 @@ export * from './TitleBar';
 export * from './TitleBarCount';
 export * from './ValidationOverview';
 export * from './ValidationPopover';
-export * from './VirtualizedValidation';
 export * from './WalkThrough';
 export * from './WalkThroughCard';
